% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoglikPen.R
\name{LoglikPen}
\alias{LoglikPen}
\title{Compute the penalized log likelihood for GLM with MIC penalty (Self-Written)}
\usage{
LoglikPen(beta, I.preselect = NULL, preselect.beta0 = FALSE, X, y, lambda,
  a, family = "gaussian")
}
\arguments{
\item{beta}{A p-dimensional vector containing the regression ceofficients.}

\item{I.preselect}{A 0-1 indicator vector of same length as \code{beta} showing preselected variables (value 0 for no penalty) that will not be penalized.}

\item{preselect.beta0}{Indicator of whether or not the intercept is pre-selected. Default is \code{FALSE}.}

\item{X}{An \eqn{n} by \eqn{p} design matrix.}

\item{y}{The \eqn{n} by 1 response vector}

\item{lambda}{The penalty parameter euqals either 2 in AIC or ln(n) in BIC (by default).
It can be specified as any value of the user's own choice.}

\item{a}{The scale parameter in the hyperbolic tangent function of the MIC penalty. By default, \eqn{a = 50}.}

\item{family}{a description of the error distribution. To use this function, \code{family} needs to be one of \code{"gaussian"},}
}
\value{
The value of the penalized log likelihood function evaluated at beta.
}
\description{
Compute the penalized log likelihood for GLM with MIC penalty (Self-Written)
}
\details{
This function is much faster than \code{\link{LoglikPenGLM}}, but it is only applicable for Gaussian linear regression, logistic regression,
and loglinear or Poisson regression models. To take advantage, sepcify  \code{family} as \code{family="gaussian"}, \code{family="binomial"}, or \code{family="poisson"} only.
}
\references{
\itemize{
\item Su, X. (2015). Variable selection via subtle uprooting.
\emph{Journal of Computational and Graphical Statistics}, \bold{24}(4): 1092--1113.
URL \url{http://www.tandfonline.com/doi/pdf/10.1080/10618600.2014.955176}
\item Su, X., Fan, J., Levine, R. A., Nunn, M. E., and Tsai, C.-L. (2016+). Sparse estimation of generalized linear
models via approximated information criteria. Submitted, \emph{Statistica Sinica}.
}
}
\seealso{
\code{\link[stats]{glm}}, \code{\link[stats]{family}}
}

