% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-glmMIC.R
\name{print.glmMIC}
\alias{print.glmMIC}
\title{The Generic \code{print} Function for Object of \code{glmMIC} Class}
\usage{
\method{print}{glmMIC}(fit, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{fit}{an object of \code{glmMIC} class.}

\item{digits}{the minimal number of significant digits. See \code{\link[base]{print.default}}.}

\item{\code{...}}{further arguments passed to or from other methods.}
}
\value{
The table of estimated regression coefficients beta and the reparameterized gamma.
}
\description{
The Generic \code{print} Function for Object of \code{glmMIC} Class
}
\details{
The (generic) print method for an \code{glmMIC} object. The results include info on the estimated gamma and beta.
Depending on the options, significance testing and confidence intervals are also provided.
}
\references{
\itemize{
\item Su, X. (2015). Variable selection via subtle uprooting.
\emph{Journal of Computational and Graphical Statistics}, \bold{24}(4): 1092--1113.
URL \url{http://www.tandfonline.com/doi/pdf/10.1080/10618600.2014.955176}
\item Su, X., Fan, J., Levine, R. A., Nunn, M. E., and Tsai, C.-L. (2016+). Sparse estimation of generalized linear
models via approximated information criteria. Submitted, \emph{Statistica Sinica}.
}
#' @seealso \code{\link[glmMIC]{glmMIC}}
}

